pyproject.toml
[project]
name = "dynamic-dataset"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, < 3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "networkx",
    "numpy",
    "Pillow",
    "reportlab>=4.0.0",
    "scipy>=1.11",
    "types-networkx>=3.6.1.20260210",
]

[dependency-groups]
dev = [
    "ruff>=0.15.4",
]
